refactor: implement Protected routes
